requirements.txt
psycopg2-binary==2.9.7
matplotlib==3.8.2
